% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_rate.R
\name{conv_rate}
\alias{conv_rate}
\title{Simulation function}
\usage{
conv_rate(seed, d, r, p1, p2, dis, gs_mean = 0, gs_sd = 10,
  unf_a = 0, unf_b = 1, dup, Nsim, linear = TRUE, cons = "vanilla",
  lambda = 1, solver = NULL)
}
\arguments{
\item{seed}{seed for generated data}

\item{d}{the vector contains different tensor size, please see "details"}

\item{r}{the vector contains different tensor rank, please see "details"}

\item{p1}{the vector contains  different number of features on mode 1}

\item{p2}{the vector contains  different number of features on mode 2}

\item{dis}{distribution of generated core tensor}

\item{gs_mean}{please see "details"}

\item{gs_sd}{please see "details"}

\item{unf_a}{please see "details"}

\item{unf_b}{please see "details"}

\item{dup}{the duplicates of simulated tensor for each ground truth}

\item{Nsim}{please see  help documents of "update_binary"}

\item{linear}{please see  help documents of "update_binary"}

\item{cons}{please see  help documents of "update_binary"}

\item{lambda}{please see  help documents of "update_binary"}

\item{solver}{please see  help documents of "update_binary"}
}
\value{
a list containing following components:

                   \code{RMSE} RMSE for different \eqn{d}

                   \code{rate} RMSE for different \eqn{d}
}
\description{
This is the function for pack of simulations using the updating function to validate the rates
}
\details{
In this scenario, for each \eqn{d_i} in vector d, we consider
              \deqn{d_1 = d_2 = d_3 = d}
             r is the same
}
