% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sele_lambda.R
\name{sele_lambda}
\alias{sele_lambda}
\title{Select penalty coefficient in constrained optimization in penalised algorithm}
\usage{
sele_lambda(seed, lambda, ...)
}
\arguments{
\item{seed}{seed for generated data}

\item{lambda}{a vector containing different penalty coefficient}

\item{...}{other passed argument(see "details")}
}
\value{
a list containing different RMSE when choosing different lambda, if we set dimension
            to be a vector, the element in list is a vector containing different RMSE correspond to
            different dimension(so is rank, p1, p2,etc.)
}
\description{
This is the function for selecting lambda in penalised optimization in supervised setting
}
\details{
other passed argument please see "conv_rate"
}
