% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sele_rank.R
\name{sele_rank}
\alias{sele_rank}
\title{Select rank of the coefficient tensor given binary tensor observation}
\usage{
sele_rank(tsr, X_covar1 = NULL, X_covar2 = NULL, rank = c(3, 5, 6,
  8), Nsim, linear = FALSE, cons = "non")
}
\arguments{
\item{tsr}{the data vector}

\item{X_covar1}{the covariate on first mode, unsupervised version please see "details"}

\item{X_covar2}{the covariate on second mode, unsupervised version please see "details"}

\item{rank}{a vector containing the several rank candidates}

\item{Nsim}{simulation times in each update}

\item{linear}{whether use linear regression to get initialization, only used when covariate is identity matrix}

\item{cons}{the constrain method, "non" for without constrain, "vanilla" for global scale down once at each iteration,

                   "penalty" for add log-barrier penalty to object function.}
}
\value{
a list containing following components:

                   \code{W1} coefficient matrix on mode-1

                   \code{W2} coefficient matrix on mode-2

                   \code{C} factor matrix on mode-3

                   \code{G} core tensor

                   \code{lglk} the lig-likelihood
}
\description{
This is the function for the selecting the rank of coefficient tensor i.e. size of the core tensor
            \eqn{B = G \times_1 W_1 \times_2 W_2 \times_3 C}
            by BIC criterion.
}
\details{
For selecting rank, recommend use non-constrain version updating.

            When you select the rank of binary in unsupervised verison, please leave two covariate to blank,
            or you can set them as identity matrix.
}
