% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_binary.R
\name{update_binary}
\alias{update_binary}
\title{Solve supervised binary tucker decomposition with covariate on two modes}
\usage{
update_binary(tsr, X_covar1 = NULL, X_covar2 = NULL, core_shape, Nsim,
  linear = TRUE, cons = "vanilla", lambda = 1, alpha = 1,
  solver = NULL)
}
\arguments{
\item{tsr}{the data vector}

\item{X_covar1}{the covariate on first mode, if you do not have covariate on mode 1 please leave it}

\item{X_covar2}{the covariate on second mode, if you do not have covariate on mode 2 please leave it}

\item{core_shape}{the tucker rank of the given tensor}

\item{Nsim}{simulation times}

\item{linear}{whether use linear regression to get initialization, only used when covariate is identity matrix}

\item{cons}{the constrain method, "non" for without constrain, "vanilla" for global scale down once at each iteration,

                   "penalty" for add log-barrier penalty to object function.}

\item{lambda}{penalty coefficient when use "penalty" constrain.}

\item{alpha}{max norm consrain on ground truth}

\item{solver}{solver for solving object function when using "penalty" constrain, see "details"}
}
\value{
a list containing following components:

                   \code{W1} coefficient matrix on mode-1

                   \code{W2} coefficient matrix on mode-2

                   \code{C} factor matrix on mode-3

                   \code{G} core tensor

                   \code{lglk} the lig-likelihood
}
\description{
This is main updating function for the update of factor matrces and  core tensor
            \eqn{B = G \times_1 W_1 \times_2 W_2 \times_3 C}
}
\details{
When you choose constrain to be "penalty". It actually add log-barrier as regularizer to
           general object function (negative log-likelihood). It use solver in function "optim" to
           solve the objective function. You can choose any solver contained in function "optim".
}
